get docker-compose tilt ready
