+++ node/docker-compose.yml
@@ -1,43 +1,51 @@
+version: "2.4"
+
+x-common-env: &common-env
+  HONEYCOMB_API_KEY: ${HONEYCOMB_API_KEY}
+  OTEL_EXPORTER_OTLP_ENDPOINT: api.honeycomb.io
+  OTEL_RESOURCE_ATTRIBUTES: app.running-in=docker
+  MESSAGE_ENDPOINT: http://message:9000
+  NAME_ENDPOINT: http://name:8000
+  YEAR_ENDPOINT: http://year:6001
+  REDIS_URL: redis
+
 services:
   frontend:
-    build: frontend
+    build:
+      context: .
+      dockerfile: ./frontend-service/Dockerfile
     environment:
-      - HONEYCOMB_API
-      - HONEYCOMB_API_KEY
-      - HONEYCOMB_DATASET
-      - NAME_ENDPOINT=name:8000
-      - MESSAGE_ENDPOINT=message:9000
-      - SERVICE_NAME=node-frontend-service
+      <<: *common-env
+      SERVICE_NAME: node-frontend-service
     ports:
       - 7777:7777
 
   message:
-    build: message-service
+    build:
+      context: .
+      dockerfile: ./message-service/Dockerfile
     environment:
-      - HONEYCOMB_API
-      - HONEYCOMB_API_KEY
-      - HONEYCOMB_DATASET
-      - SERVICE_NAME=node-message-service
+      <<: *common-env
+      SERVICE_NAME: node-message-service
     ports:
       - 9000:9000
 
   name:
-    build: name-service
+    build:
+      context: .
+      dockerfile: ./name-service/Dockerfile
     environment:
-      - HONEYCOMB_API
-      - HONEYCOMB_API_KEY
-      - HONEYCOMB_DATASET
-      - SERVICE_NAME=node-name-service
-      - YEAR_ENDPOINT=year:6001
+      <<: *common-env
+      SERVICE_NAME: node-name-service
     ports:
       - 8000:8000
 
   year:
-    build: year-service
+    build:
+      context: .
+      dockerfile: ./year-service/Dockerfile
     environment:
-      - HONEYCOMB_API_KEY
-      - HONEYCOMB_DATASET
-      - OTLP_ENDPOINT
-      - SERVICE_NAME=node-year-service
+      <<: *common-env
+      SERVICE_NAME: node-year-service
     ports:
       - 6001:6001
