+++ node/frontend-service/.dockerignore
