+++ node/year-service/Dockerfile
@@ -6,14 +6,14 @@ WORKDIR /usr/src/app
 # Install app dependencies
 # A wildcard is used to ensure both package.json AND package-lock.json are copied
 # where available (npm@5+)
-COPY package*.json ./
+COPY year-service/package*.json ./
 
 RUN npm install
 # If you are building your code for production
 # RUN npm ci --only=production
 
 # Bundle app source
-COPY . .
+COPY year-service/. .
 
 EXPOSE 6001
 CMD [ "npm", "start" ]
